% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bfit.h.R
\name{bfit}
\alias{bfit}
\title{Bootstrap Item Fit}
\usage{
bfit(
  data,
  vars,
  step = 1,
  bn = 100,
  binfit = TRUE,
  boutfit = FALSE,
  inplot = FALSE,
  outplot = FALSE,
  angle = 0
)
}
\arguments{
\item{data}{The data as a data frame.}

\item{vars}{.}

\item{step}{.}

\item{bn}{.}

\item{binfit}{.}

\item{boutfit}{.}

\item{inplot}{.}

\item{outplot}{.}

\item{angle}{.}
}
\value{
A results object containing:
\tabular{llllll}{
  \code{results$instructions} \tab \tab \tab \tab \tab a html \cr
  \code{results$item$binfit} \tab \tab \tab \tab \tab a table \cr
  \code{results$item$boutfit} \tab \tab \tab \tab \tab a table \cr
  \code{results$inplot} \tab \tab \tab \tab \tab an image \cr
  \code{results$outplot} \tab \tab \tab \tab \tab an image \cr
}
}
\description{
Bootstrap Item Fit
}
