% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dif.h.R
\name{dif}
\alias{dif}
\title{Differential Item Functioning}
\usage{
dif(
  data,
  vars,
  facs,
  model = "dicho",
  z = FALSE,
  plot1 = FALSE,
  comp = FALSE,
  plot2 = FALSE,
  angle = 0,
  plot3 = FALSE,
  plot4 = FALSE,
  plot5 = FALSE,
  comp1 = FALSE,
  z1 = FALSE,
  plot6 = FALSE,
  angle1 = 0,
  plot7 = FALSE,
  plot8 = FALSE
)
}
\arguments{
\item{data}{The data as a data frame.}

\item{vars}{.}

\item{facs}{.}

\item{model}{.}

\item{z}{.}

\item{plot1}{.}

\item{comp}{.}

\item{plot2}{.}

\item{angle}{.}

\item{plot3}{.}

\item{plot4}{.}

\item{plot5}{.}

\item{comp1}{.}

\item{z1}{.}

\item{plot6}{.}

\item{angle1}{.}

\item{plot7}{.}

\item{plot8}{.}
}
\value{
A results object containing:
\tabular{llllll}{
  \code{results$instructions} \tab \tab \tab \tab \tab a html \cr
  \code{results$text} \tab \tab \tab \tab \tab a preformatted \cr
  \code{results$comp} \tab \tab \tab \tab \tab a table \cr
  \code{results$z} \tab \tab \tab \tab \tab a table \cr
  \code{results$plot1} \tab \tab \tab \tab \tab an image \cr
  \code{results$plot2} \tab \tab \tab \tab \tab an image \cr
  \code{results$plot3} \tab \tab \tab \tab \tab an image \cr
  \code{results$plot4} \tab \tab \tab \tab \tab an image \cr
  \code{results$comp1} \tab \tab \tab \tab \tab a table \cr
  \code{results$z1} \tab \tab \tab \tab \tab a table \cr
  \code{results$plot5} \tab \tab \tab \tab \tab an image \cr
  \code{results$plot6} \tab \tab \tab \tab \tab an image \cr
  \code{results$plot7} \tab \tab \tab \tab \tab an image \cr
  \code{results$plot8} \tab \tab \tab \tab \tab an image \cr
}

Tables can be converted to data frames with \code{asDF} or \code{\link{as.data.frame}}. For example:

\code{results$comp$asDF}

\code{as.data.frame(results$comp)}
}
\description{
Differential Item Functioning
}
