% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/equating.h.R
\name{equating}
\alias{equating}
\title{Linear Equating}
\usage{
equating(
  data,
  ind,
  dep,
  contabx = FALSE,
  contabxy = FALSE,
  lineq = TRUE,
  cdfplot = FALSE,
  escore = FALSE
)
}
\arguments{
\item{data}{.}

\item{ind}{.}

\item{dep}{.}

\item{contabx}{.}

\item{contabxy}{.}

\item{lineq}{.}

\item{cdfplot}{.}

\item{escore}{.}
}
\value{
A results object containing:
\tabular{llllll}{
  \code{results$instructions} \tab \tab \tab \tab \tab a html \cr
  \code{results$lineq} \tab \tab \tab \tab \tab a table \cr
}

Tables can be converted to data frames with \code{asDF} or \code{\link{as.data.frame}}. For example:

\code{results$lineq$asDF}

\code{as.data.frame(results$lineq)}
}
\description{
Linear Equating
}
