% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lcgm.h.R
\name{lcgm}
\alias{lcgm}
\title{Latent Class Growth Modeling}
\usage{
lcgm(
  data,
  vars,
  model = list(),
  nc = 2,
  thr = "FALSE",
  fit = FALSE,
  est = FALSE,
  desc = TRUE,
  cp = FALSE,
  plot1 = FALSE,
  plot = FALSE,
  raw = "FALSE",
  miss = "listwise",
  plot2 = FALSE,
  auxVar,
  use3step = FALSE
)
}
\arguments{
\item{data}{.}

\item{vars}{.}

\item{model}{.}

\item{nc}{.}

\item{thr}{.}

\item{fit}{.}

\item{est}{.}

\item{desc}{.}

\item{cp}{.}

\item{plot1}{.}

\item{plot}{.}

\item{raw}{.}

\item{miss}{.}

\item{plot2}{.}

\item{auxVar}{.}

\item{use3step}{.}
}
\value{
A results object containing:
\tabular{llllll}{
  \code{results$instructions} \tab \tab \tab \tab \tab a html \cr
  \code{results$text} \tab \tab \tab \tab \tab a preformatted \cr
  \code{results$progressBarHTML} \tab \tab \tab \tab \tab a html \cr
  \code{results$desc} \tab \tab \tab \tab \tab a table \cr
  \code{results$fit} \tab \tab \tab \tab \tab a table \cr
  \code{results$est} \tab \tab \tab \tab \tab a table \cr
  \code{results$cp} \tab \tab \tab \tab \tab a table \cr
  \code{results$mem} \tab \tab \tab \tab \tab an output \cr
  \code{results$plot2} \tab \tab \tab \tab \tab an image \cr
  \code{results$plot1} \tab \tab \tab \tab \tab an image \cr
  \code{results$plot} \tab \tab \tab \tab \tab an image \cr
  \code{results$threeStepSummary} \tab \tab \tab \tab \tab a table \cr
  \code{results$threeStepClassStats} \tab \tab \tab \tab \tab a table \cr
  \code{results$threeStepPairwise} \tab \tab \tab \tab \tab a table \cr
}

Tables can be converted to data frames with \code{asDF} or \code{\link{as.data.frame}}. For example:

\code{results$desc$asDF}

\code{as.data.frame(results$desc)}
}
\description{
Latent Class Growth Modeling
}
