% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linking.h.R
\name{linking}
\alias{linking}
\title{Composite Linking}
\usage{
linking(
  data,
  ind,
  dep,
  method = "linear",
  plot = FALSE,
  width = 500,
  height = 500
)
}
\arguments{
\item{data}{.}

\item{ind}{.}

\item{dep}{.}

\item{method}{.}

\item{plot}{.}

\item{width}{.}

\item{height}{.}
}
\value{
A results object containing:
\tabular{llllll}{
  \code{results$instructions} \tab \tab \tab \tab \tab a html \cr
  \code{results$con} \tab \tab \tab \tab \tab a table \cr
  \code{results$plot} \tab \tab \tab \tab \tab an image \cr
}

Tables can be converted to data frames with \code{asDF} or \code{\link{as.data.frame}}. For example:

\code{results$con$asDF}

\code{as.data.frame(results$con)}
}
\description{
Composite Linking
}
