% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lltm.h.R
\name{lltm}
\alias{lltm}
\title{Linear Logisitic Test Model}
\usage{
lltm(
  data,
  vars,
  mat = "1,0,1,1,1,0,1,0,1,0,1,1,1,1,1,0,0,0,1,1,0,1,0,1,1,  -0.5, 1,  1, -1,0",
  col = 3,
  items = TRUE,
  lr = TRUE,
  ml = TRUE,
  wald = FALSE,
  eta = FALSE,
  beta = FALSE,
  plot = FALSE,
  comp = FALSE,
  plot1 = FALSE,
  con = FALSE,
  wdiag = FALSE,
  resid = FALSE,
  plot2 = FALSE
)
}
\arguments{
\item{data}{The data as a data frame.}

\item{vars}{.}

\item{mat}{.}

\item{col}{.}

\item{items}{.}

\item{lr}{.}

\item{ml}{.}

\item{wald}{.}

\item{eta}{.}

\item{beta}{.}

\item{plot}{.}

\item{comp}{.}

\item{plot1}{.}

\item{con}{.}

\item{wdiag}{.}

\item{resid}{.}

\item{plot2}{.}
}
\value{
A results object containing:
\tabular{llllll}{
  \code{results$instructions} \tab \tab \tab \tab \tab a html \cr
  \code{results$text} \tab \tab \tab \tab \tab a preformatted \cr
  \code{results$ra$items} \tab \tab \tab \tab \tab a table \cr
  \code{results$ra$lr} \tab \tab \tab \tab \tab a table \cr
  \code{results$ra$ml} \tab \tab \tab \tab \tab a table \cr
  \code{results$ra$wald} \tab \tab \tab \tab \tab a table \cr
  \code{results$ll$eta} \tab \tab \tab \tab \tab a table \cr
  \code{results$ll$beta} \tab \tab \tab \tab \tab a table \cr
  \code{results$ll$comp} \tab \tab \tab \tab \tab a table \cr
  \code{results$con} \tab \tab \tab \tab \tab a table \cr
  \code{results$wdiag} \tab \tab \tab \tab \tab a table \cr
  \code{results$plot} \tab \tab \tab \tab \tab an image \cr
  \code{results$plot1} \tab \tab \tab \tab \tab an image \cr
  \code{results$plot2} \tab \tab \tab \tab \tab an image \cr
  \code{results$resid} \tab \tab \tab \tab \tab a table \cr
}

Tables can be converted to data frames with \code{asDF} or \code{\link{as.data.frame}}. For example:

\code{results$con$asDF}

\code{as.data.frame(results$con)}
}
\description{
Linear Logisitic Test Model
}
