% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lpa.h.R
\name{lpa}
\alias{lpa}
\title{Latent Profile Analysis}
\usage{
lpa(
  data,
  vars,
  run,
  nc = 2,
  variances = "equal",
  covariances = "zero",
  overall = TRUE,
  fit = FALSE,
  est = FALSE,
  plot = FALSE,
  plot1 = FALSE,
  plot3 = FALSE,
  plot2 = FALSE,
  plot4 = FALSE,
  plot5 = FALSE,
  angle = 0,
  line = "FALSE",
  use3step = FALSE,
  auxVar
)
}
\arguments{
\item{data}{.}

\item{vars}{.}

\item{run}{.}

\item{nc}{.}

\item{variances}{.}

\item{covariances}{.}

\item{overall}{.}

\item{fit}{.}

\item{est}{.}

\item{plot}{.}

\item{plot1}{.}

\item{plot3}{.}

\item{plot2}{.}

\item{plot4}{.}

\item{plot5}{.}

\item{angle}{a number from 0 to 90 defining the angle of the x-axis labels,
where 0 degrees represents completely horizontal labels.}

\item{line}{.}

\item{use3step}{.}

\item{auxVar}{.}
}
\value{
A results object containing:
\tabular{llllll}{
  \code{results$instructions} \tab \tab \tab \tab \tab a html \cr
  \code{results$progressBarHTML} \tab \tab \tab \tab \tab a html \cr
  \code{results$text} \tab \tab \tab \tab \tab a preformatted \cr
  \code{results$overall} \tab \tab \tab \tab \tab a table \cr
  \code{results$fit} \tab \tab \tab \tab \tab a table \cr
  \code{results$est} \tab \tab \tab \tab \tab a table \cr
  \code{results$pc} \tab \tab \tab \tab \tab an output \cr
  \code{results$post} \tab \tab \tab \tab \tab an output \cr
  \code{results$plot} \tab \tab \tab \tab \tab an image \cr
  \code{results$plot3} \tab \tab \tab \tab \tab an image \cr
  \code{results$plot2} \tab \tab \tab \tab \tab an image \cr
  \code{results$plot1} \tab \tab \tab \tab \tab an image \cr
  \code{results$plot4} \tab \tab \tab \tab \tab an image \cr
  \code{results$plot5} \tab \tab \tab \tab \tab an image \cr
  \code{results$lpa3_means} \tab \tab \tab \tab \tab a table \cr
  \code{results$lpa3_omnibus} \tab \tab \tab \tab \tab a table \cr
  \code{results$lpa3_pairwise} \tab \tab \tab \tab \tab a table \cr
}

Tables can be converted to data frames with \code{asDF} or \code{\link{as.data.frame}}. For example:

\code{results$overall$asDF}

\code{as.data.frame(results$overall)}
}
\description{
Latent Profile Analysis
}
