% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixture.h.R
\name{mixture}
\alias{mixture}
\title{Rasch Mixture Model}
\usage{
mixture(
  data,
  vars,
  nc = 2,
  step = 1,
  type = "RSM",
  aic = TRUE,
  bic = TRUE,
  caic = TRUE,
  imean = FALSE,
  imeasure = FALSE,
  ise = FALSE,
  infit = FALSE,
  outfit = FALSE,
  pbis = FALSE,
  average = FALSE,
  iplot = TRUE,
  wrightmap = FALSE
)
}
\arguments{
\item{data}{The data as a data frame.}

\item{vars}{.}

\item{nc}{.}

\item{step}{.}

\item{type}{.}

\item{aic}{.}

\item{bic}{.}

\item{caic}{.}

\item{imean}{.}

\item{imeasure}{.}

\item{ise}{.}

\item{infit}{.}

\item{outfit}{.}

\item{pbis}{.}

\item{average}{.}

\item{iplot}{.}

\item{wrightmap}{.}
}
\value{
A results object containing:
\tabular{llllll}{
  \code{results$instructions} \tab \tab \tab \tab \tab a html \cr
  \code{results$item$model} \tab \tab \tab \tab \tab a table \cr
  \code{results$item$imean} \tab \tab \tab \tab \tab a table \cr
  \code{results$item$imeasure} \tab \tab \tab \tab \tab a table \cr
  \code{results$item$ise} \tab \tab \tab \tab \tab a table \cr
  \code{results$item$infit} \tab \tab \tab \tab \tab a table \cr
  \code{results$item$outfit} \tab \tab \tab \tab \tab a table \cr
  \code{results$item$pbis} \tab \tab \tab \tab \tab a table \cr
  \code{results$person$average} \tab \tab \tab \tab \tab a table \cr
  \code{results$iplot} \tab \tab \tab \tab \tab an image \cr
  \code{results$plot} \tab \tab \tab \tab \tab an image \cr
  \code{results$pclass} \tab \tab \tab \tab \tab an output \cr
}
}
\description{
Rasch Mixture Model
}
