% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pfit.h.R
\name{pfit}
\alias{pfit}
\title{Bootstrap item fit with P values}
\usage{
pfit(
  data,
  vars,
  type = "bi",
  adj = "BH",
  bn = 100,
  outfit = TRUE,
  infit = FALSE
)
}
\arguments{
\item{data}{The data as a data frame.}

\item{vars}{.}

\item{type}{.}

\item{adj}{.}

\item{bn}{.}

\item{outfit}{.}

\item{infit}{.}
}
\value{
A results object containing:
\tabular{llllll}{
  \code{results$instructions} \tab \tab \tab \tab \tab a html \cr
  \code{results$outfit} \tab \tab \tab \tab \tab a table \cr
  \code{results$infit} \tab \tab \tab \tab \tab a table \cr
}

Tables can be converted to data frames with \code{asDF} or \code{\link{as.data.frame}}. For example:

\code{results$outfit$asDF}

\code{as.data.frame(results$outfit)}
}
\description{
Bootstrap item fit with P values
}
