% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bfit.h.R
\name{bfit}
\alias{bfit}
\title{Bootstrap Item Fit}
\usage{
bfit(
  data,
  mode = "simple",
  vars,
  vars1,
  step = 1,
  bn = 100,
  binfit = TRUE,
  boutfit = FALSE,
  inplot = FALSE,
  outplot = FALSE,
  angle = 0,
  type = "bi",
  co = "yes",
  adj,
  bn1 = 100,
  outfit = FALSE,
  infit = FALSE,
  noutfit = TRUE,
  ninfit = FALSE,
  width = 500,
  height = 500
)
}
\arguments{
\item{data}{The data as a data frame.}

\item{mode}{.}

\item{vars}{.}

\item{vars1}{.}

\item{step}{.}

\item{bn}{.}

\item{binfit}{.}

\item{boutfit}{.}

\item{inplot}{.}

\item{outplot}{.}

\item{angle}{.}

\item{type}{.}

\item{co}{.}

\item{adj}{.}

\item{bn1}{.}

\item{outfit}{.}

\item{infit}{.}

\item{noutfit}{.}

\item{ninfit}{.}

\item{width}{.}

\item{height}{.}
}
\value{
A results object containing:
\tabular{llllll}{
  \code{results$instructions} \tab \tab \tab \tab \tab a html \cr
  \code{results$item$binfit} \tab \tab \tab \tab \tab a table \cr
  \code{results$item$boutfit} \tab \tab \tab \tab \tab a table \cr
  \code{results$inplot} \tab \tab \tab \tab \tab an image \cr
  \code{results$outplot} \tab \tab \tab \tab \tab an image \cr
  \code{results$instructions1} \tab \tab \tab \tab \tab a html \cr
  \code{results$noutfit} \tab \tab \tab \tab \tab a table \cr
  \code{results$ninfit} \tab \tab \tab \tab \tab a table \cr
  \code{results$outfit} \tab \tab \tab \tab \tab a table \cr
  \code{results$infit} \tab \tab \tab \tab \tab a table \cr
}

Tables can be converted to data frames with \code{asDF} or \code{\link{as.data.frame}}. For example:

\code{results$noutfit$asDF}

\code{as.data.frame(results$noutfit)}
}
\description{
Bootstrap Item Fit
}
