% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bfit.h.R
\name{bfit}
\alias{bfit}
\title{Bootstrap Item Fit}
\usage{
bfit(
  data,
  vars,
  step = 1,
  type = "RSM",
  bn = 1000,
  binfit = TRUE,
  boutfit = FALSE
)
}
\arguments{
\item{data}{The data as a data frame.}

\item{vars}{.}

\item{step}{.}

\item{type}{.}

\item{bn}{.}

\item{binfit}{.}

\item{boutfit}{.}
}
\value{
A results object containing:
\tabular{llllll}{
  \code{results$instructions} \tab \tab \tab \tab \tab a html \cr
  \code{results$item$binfit} \tab \tab \tab \tab \tab a table \cr
  \code{results$item$boutfit} \tab \tab \tab \tab \tab a table \cr
}
}
\description{
Bootstrap Item Fit
}
