% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bfit.h.R
\name{bfit}
\alias{bfit}
\title{Bootstrap Item Fit}
\usage{
bfit(
  data,
  vars,
  step = 1,
  type = "RSM",
  bn = 1000,
  bfit = TRUE,
  inplot = FALSE,
  outplot = FALSE
)
}
\arguments{
\item{data}{The data as a data frame.}

\item{vars}{.}

\item{step}{.}

\item{type}{.}

\item{bn}{.}

\item{bfit}{.}

\item{inplot}{.}

\item{outplot}{.}
}
\value{
A results object containing:
\tabular{llllll}{
  \code{results$instructions} \tab \tab \tab \tab \tab a html \cr
  \code{results$bfit} \tab \tab \tab \tab \tab a table \cr
}

Tables can be converted to data frames with \code{asDF} or \code{\link{as.data.frame}}. For example:

\code{results$bfit$asDF}

\code{as.data.frame(results$bfit)}
}
\description{
Bootstrap Item Fit
}
