% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dif.h.R
\name{dif}
\alias{dif}
\title{Differential Item Functioning}
\usage{
dif(data, vars, facs, z = TRUE)
}
\arguments{
\item{data}{The data as a data frame.}

\item{vars}{.}

\item{facs}{.}

\item{z}{.}
}
\value{
A results object containing:
\tabular{llllll}{
  \code{results$instructions} \tab \tab \tab \tab \tab a html \cr
  \code{results$text} \tab \tab \tab \tab \tab a preformatted \cr
  \code{results$z} \tab \tab \tab \tab \tab a table \cr
}

Tables can be converted to data frames with \code{asDF} or \code{\link{as.data.frame}}. For example:

\code{results$z$asDF}

\code{as.data.frame(results$z)}
}
\description{
Differential Item Functioning
}
