% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dif.h.R
\name{dif}
\alias{dif}
\title{Differential Item Functioning}
\usage{
dif(
  data,
  vars,
  facs,
  z = TRUE,
  plot1 = FALSE,
  width1 = 500,
  height1 = 500,
  comp = FALSE,
  plot2 = FALSE,
  angle = 0,
  width2 = 500,
  height2 = 500,
  plot3 = FALSE,
  ci = 95,
  width3 = 500,
  height3 = 500
)
}
\arguments{
\item{data}{The data as a data frame.}

\item{vars}{.}

\item{facs}{.}

\item{z}{.}

\item{plot1}{.}

\item{width1}{.}

\item{height1}{.}

\item{comp}{.}

\item{plot2}{.}

\item{angle}{.}

\item{width2}{.}

\item{height2}{.}

\item{plot3}{.}

\item{ci}{.}

\item{width3}{.}

\item{height3}{.}
}
\value{
A results object containing:
\tabular{llllll}{
  \code{results$instructions} \tab \tab \tab \tab \tab a html \cr
  \code{results$text} \tab \tab \tab \tab \tab a preformatted \cr
  \code{results$z} \tab \tab \tab \tab \tab a table \cr
  \code{results$plot1} \tab \tab \tab \tab \tab an image \cr
  \code{results$comp} \tab \tab \tab \tab \tab a table \cr
  \code{results$plot2} \tab \tab \tab \tab \tab an image \cr
  \code{results$plot3} \tab \tab \tab \tab \tab an image \cr
}

Tables can be converted to data frames with \code{asDF} or \code{\link{as.data.frame}}. For example:

\code{results$z$asDF}

\code{as.data.frame(results$z)}
}
\description{
Differential Item Functioning
}
