% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dif.h.R
\name{dif}
\alias{dif}
\title{Differential Item Functioning}
\usage{
dif(
  data,
  vars,
  facs,
  model = "dicho",
  z = TRUE,
  plot1 = FALSE,
  width1 = 500,
  height1 = 500,
  comp = FALSE,
  plot2 = FALSE,
  angle = 0,
  width2 = 500,
  height2 = 500,
  plot3 = FALSE,
  width3 = 500,
  height3 = 500,
  plot4 = FALSE,
  width4 = 500,
  height4 = 500,
  plot5 = FALSE,
  width5 = 500,
  height5 = 500,
  comp1 = FALSE,
  plot6 = FALSE,
  angle1 = 0,
  width6 = 500,
  height6 = 500
)
}
\arguments{
\item{data}{The data as a data frame.}

\item{vars}{.}

\item{facs}{.}

\item{model}{.}

\item{z}{.}

\item{plot1}{.}

\item{width1}{.}

\item{height1}{.}

\item{comp}{.}

\item{plot2}{.}

\item{angle}{.}

\item{width2}{.}

\item{height2}{.}

\item{plot3}{.}

\item{width3}{.}

\item{height3}{.}

\item{plot4}{.}

\item{width4}{.}

\item{height4}{.}

\item{plot5}{.}

\item{width5}{.}

\item{height5}{.}

\item{comp1}{.}

\item{plot6}{.}

\item{angle1}{.}

\item{width6}{.}

\item{height6}{.}
}
\value{
A results object containing:
\tabular{llllll}{
  \code{results$instructions} \tab \tab \tab \tab \tab a html \cr
  \code{results$text} \tab \tab \tab \tab \tab a preformatted \cr
  \code{results$z} \tab \tab \tab \tab \tab a table \cr
  \code{results$plot1} \tab \tab \tab \tab \tab an image \cr
  \code{results$comp} \tab \tab \tab \tab \tab a table \cr
  \code{results$plot2} \tab \tab \tab \tab \tab an image \cr
  \code{results$plot3} \tab \tab \tab \tab \tab an image \cr
  \code{results$plot4} \tab \tab \tab \tab \tab an image \cr
  \code{results$plot5} \tab \tab \tab \tab \tab an image \cr
  \code{results$comp1} \tab \tab \tab \tab \tab a table \cr
  \code{results$plot6} \tab \tab \tab \tab \tab an image \cr
}

Tables can be converted to data frames with \code{asDF} or \code{\link{as.data.frame}}. For example:

\code{results$z$asDF}

\code{as.data.frame(results$z)}
}
\description{
Differential Item Functioning
}
