% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/equating.h.R
\name{equating}
\alias{equating}
\title{CTT Equating}
\usage{
equating(
  mode = "simple",
  data,
  ind,
  dep,
  lineq = FALSE,
  contabx = FALSE,
  contaby = FALSE,
  plot = FALSE,
  ind1,
  dep1,
  design = "single",
  con = FALSE,
  contabx1 = FALSE,
  contaby1 = FALSE,
  plot1 = FALSE,
  width = 500,
  height = 500,
  width1 = 500,
  height1 = 500
)
}
\arguments{
\item{mode}{.}

\item{data}{.}

\item{ind}{.}

\item{dep}{.}

\item{lineq}{.}

\item{contabx}{.}

\item{contaby}{.}

\item{plot}{.}

\item{ind1}{.}

\item{dep1}{.}

\item{design}{.}

\item{con}{.}

\item{contabx1}{.}

\item{contaby1}{.}

\item{plot1}{.}

\item{width}{.}

\item{height}{.}

\item{width1}{.}

\item{height1}{.}
}
\value{
A results object containing:
\tabular{llllll}{
  \code{results$instructions} \tab \tab \tab \tab \tab a html \cr
  \code{results$lineq} \tab \tab \tab \tab \tab a table \cr
  \code{results$contabx} \tab \tab \tab \tab \tab a table \cr
  \code{results$contaby} \tab \tab \tab \tab \tab a table \cr
  \code{results$plot} \tab \tab \tab \tab \tab an image \cr
  \code{results$escore} \tab \tab \tab \tab \tab an output \cr
  \code{results$instructions1} \tab \tab \tab \tab \tab a html \cr
  \code{results$con} \tab \tab \tab \tab \tab a table \cr
  \code{results$contabx1} \tab \tab \tab \tab \tab a table \cr
  \code{results$contaby1} \tab \tab \tab \tab \tab a table \cr
  \code{results$escore1} \tab \tab \tab \tab \tab an output \cr
  \code{results$plot1} \tab \tab \tab \tab \tab an image \cr
}

Tables can be converted to data frames with \code{asDF} or \code{\link{as.data.frame}}. For example:

\code{results$lineq$asDF}

\code{as.data.frame(results$lineq)}
}
\description{
CTT Equating
}
