% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/equating.h.R
\name{equating}
\alias{equating}
\title{Linear Equating}
\usage{
equating(
  data,
  ind,
  dep,
  contabx = TRUE,
  contabxy = FALSE,
  lineq = FALSE,
  cdfplot = FALSE,
  escore = FALSE
)
}
\arguments{
\item{data}{.}

\item{ind}{.}

\item{dep}{.}

\item{contabx}{.}

\item{contabxy}{.}

\item{lineq}{.}

\item{cdfplot}{.}

\item{escore}{.}
}
\value{
A results object containing:
\tabular{llllll}{
  \code{results$instructions} \tab \tab \tab \tab \tab a html \cr
}
}
\description{
Linear Equating
}
