% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/equi.h.R
\name{equi}
\alias{equi}
\title{Equipercentile Equating}
\usage{
equi(data, ind, dep, design, con = TRUE, plot = FALSE)
}
\arguments{
\item{data}{.}

\item{ind}{.}

\item{dep}{.}

\item{design}{.}

\item{con}{.}

\item{plot}{.}
}
\value{
A results object containing:
\tabular{llllll}{
  \code{results$instructions} \tab \tab \tab \tab \tab a html \cr
  \code{results$con} \tab \tab \tab \tab \tab a table \cr
  \code{results$escore} \tab \tab \tab \tab \tab an output \cr
  \code{results$plot} \tab \tab \tab \tab \tab an image \cr
}

Tables can be converted to data frames with \code{asDF} or \code{\link{as.data.frame}}. For example:

\code{results$con$asDF}

\code{as.data.frame(results$con)}
}
\description{
Equipercentile Equating
}
