% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lca.h.R
\name{lca}
\alias{lca}
\title{Latent Class Analysis}
\usage{
lca(
  data,
  vars,
  covs,
  nc = 2,
  fit = TRUE,
  comp = FALSE,
  cp = FALSE,
  ip = FALSE,
  cf = FALSE,
  coef = FALSE,
  r = FALSE,
  plot = FALSE,
  plot1 = FALSE,
  plot2 = FALSE,
  angle = 0,
  plot3 = FALSE,
  width = 500,
  height = 500,
  width1 = 500,
  height1 = 500,
  width2 = 500,
  height2 = 500,
  width3 = 500,
  height3 = 500
)
}
\arguments{
\item{data}{The data as a data frame.}

\item{vars}{.}

\item{covs}{.}

\item{nc}{.}

\item{fit}{.}

\item{comp}{.}

\item{cp}{.}

\item{ip}{.}

\item{cf}{.}

\item{coef}{.}

\item{r}{.}

\item{plot}{.}

\item{plot1}{.}

\item{plot2}{.}

\item{angle}{a number from 0 to 90 defining the angle of the x-axis labels,
where 0 degrees represents completely horizontal labels.}

\item{plot3}{.}

\item{width}{.}

\item{height}{.}

\item{width1}{.}

\item{height1}{.}

\item{width2}{.}

\item{height2}{.}

\item{width3}{.}

\item{height3}{.}
}
\value{
A results object containing:
\tabular{llllll}{
  \code{results$instructions} \tab \tab \tab \tab \tab a html \cr
  \code{results$mf$fit} \tab \tab \tab \tab \tab a table \cr
  \code{results$mf$comp} \tab \tab \tab \tab \tab a table \cr
  \code{results$ip} \tab \tab \tab \tab \tab an array of tables \cr
  \code{results$pro$cp} \tab \tab \tab \tab \tab a table \cr
  \code{results$pro$cf} \tab \tab \tab \tab \tab a table \cr
  \code{results$text} \tab \tab \tab \tab \tab a preformatted \cr
  \code{results$cm} \tab \tab \tab \tab \tab an output \cr
  \code{results$plot3} \tab \tab \tab \tab \tab an image \cr
  \code{results$plot} \tab \tab \tab \tab \tab an image \cr
  \code{results$plot1} \tab \tab \tab \tab \tab an image \cr
  \code{results$plot2} \tab \tab \tab \tab \tab an image \cr
  \code{results$r} \tab \tab \tab \tab \tab a preformatted \cr
}
}
\description{
Latent Class Analysis
}
