% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lca.h.R
\name{lca}
\alias{lca}
\title{Latent Class Analysis}
\usage{
lca(data, vars, nc = 2, fit = TRUE, cp = TRUE, ip = FALSE, plot = FALSE)
}
\arguments{
\item{data}{The data as a data frame.}

\item{vars}{.}

\item{nc}{.}

\item{fit}{.}

\item{cp}{.}

\item{ip}{.}

\item{plot}{.}
}
\value{
A results object containing:
\tabular{llllll}{
  \code{results$instructions} \tab \tab \tab \tab \tab a html \cr
  \code{results$fit} \tab \tab \tab \tab \tab a table \cr
  \code{results$cp} \tab \tab \tab \tab \tab a table \cr
  \code{results$ip} \tab \tab \tab \tab \tab a preformatted \cr
  \code{results$cm} \tab \tab \tab \tab \tab an output \cr
  \code{results$plot} \tab \tab \tab \tab \tab an image \cr
}

Tables can be converted to data frames with \code{asDF} or \code{\link{as.data.frame}}. For example:

\code{results$fit$asDF}

\code{as.data.frame(results$fit)}
}
\description{
Latent Class Analysis
}
