% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lcaord.h.R
\name{lcaord}
\alias{lcaord}
\title{LCA with ordinal indicators}
\usage{
lcaord(
  data,
  vars,
  nc = 2,
  desc = TRUE,
  fit = FALSE,
  cp = FALSE,
  plot = FALSE,
  width = 500,
  height = 500,
  plot1 = FALSE,
  width1 = 500,
  height1 = 500,
  angle = 0,
  miss = "fiml"
)
}
\arguments{
\item{data}{.}

\item{vars}{.}

\item{nc}{.}

\item{desc}{.}

\item{fit}{.}

\item{cp}{.}

\item{plot}{.}

\item{width}{.}

\item{height}{.}

\item{plot1}{.}

\item{width1}{.}

\item{height1}{.}

\item{angle}{a number from 0 to 90 defining the angle of the x-axis labels,
where 0 degrees represents completely horizontal labels.}

\item{miss}{.}
}
\value{
A results object containing:
\tabular{llllll}{
  \code{results$instructions} \tab \tab \tab \tab \tab a html \cr
  \code{results$text} \tab \tab \tab \tab \tab a preformatted \cr
  \code{results$progressBarHTML} \tab \tab \tab \tab \tab a html \cr
  \code{results$desc} \tab \tab \tab \tab \tab a table \cr
  \code{results$fit} \tab \tab \tab \tab \tab a table \cr
  \code{results$cp} \tab \tab \tab \tab \tab a table \cr
  \code{results$mem} \tab \tab \tab \tab \tab an output \cr
  \code{results$plot1} \tab \tab \tab \tab \tab an image \cr
  \code{results$plot} \tab \tab \tab \tab \tab an image \cr
}

Tables can be converted to data frames with \code{asDF} or \code{\link{as.data.frame}}. For example:

\code{results$desc$asDF}

\code{as.data.frame(results$desc)}
}
\description{
LCA with ordinal indicators
}
