% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lcgm.h.R
\name{lcgm}
\alias{lcgm}
\title{Latent Class Growth Modeling}
\usage{
lcgm(
  data,
  vars,
  model = list(),
  nc = 2,
  type = "conti",
  variance = "equal",
  fit = TRUE,
  plot = FALSE,
  raw = "FALSE"
)
}
\arguments{
\item{data}{.}

\item{vars}{.}

\item{model}{.}

\item{nc}{.}

\item{type}{.}

\item{variance}{.}

\item{fit}{.}

\item{plot}{.}

\item{raw}{.}
}
\value{
A results object containing:
\tabular{llllll}{
  \code{results$text} \tab \tab \tab \tab \tab a preformatted \cr
  \code{results$fit} \tab \tab \tab \tab \tab a table \cr
  \code{results$plot} \tab \tab \tab \tab \tab an image \cr
}

Tables can be converted to data frames with \code{asDF} or \code{\link{as.data.frame}}. For example:

\code{results$fit$asDF}

\code{as.data.frame(results$fit)}
}
\description{
Latent Class Growth Modeling
}
