% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lcgm.h.R
\name{lcgm}
\alias{lcgm}
\title{Latent Class Growth Modeling}
\usage{
lcgm(data, vars, model = list(), fit = TRUE)
}
\arguments{
\item{data}{.}

\item{vars}{.}

\item{model}{.}

\item{fit}{.}
}
\value{
A results object containing:
\tabular{llllll}{
  \code{results$text} \tab \tab \tab \tab \tab a preformatted \cr
}
}
\description{
Latent Class Growth Modeling
}
