% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lcgm.h.R
\name{lcgm}
\alias{lcgm}
\title{Latent Class Growth Modeling}
\usage{
lcgm(
  data,
  vars,
  model = list(),
  nc = 2,
  thr = "FALSE",
  mc = TRUE,
  fit = FALSE,
  est = FALSE,
  desc = FALSE,
  cp = FALSE,
  plot1 = FALSE,
  plot = FALSE,
  raw = "FALSE",
  width = 500,
  height = 500,
  width1 = 500,
  height1 = 500,
  miss = "fiml"
)
}
\arguments{
\item{data}{.}

\item{vars}{.}

\item{model}{.}

\item{nc}{.}

\item{thr}{.}

\item{mc}{.}

\item{fit}{.}

\item{est}{.}

\item{desc}{.}

\item{cp}{.}

\item{plot1}{.}

\item{plot}{.}

\item{raw}{.}

\item{width}{.}

\item{height}{.}

\item{width1}{.}

\item{height1}{.}

\item{miss}{.}
}
\value{
A results object containing:
\tabular{llllll}{
  \code{results$instructions} \tab \tab \tab \tab \tab a html \cr
  \code{results$text} \tab \tab \tab \tab \tab a preformatted \cr
  \code{results$mc} \tab \tab \tab \tab \tab a table \cr
  \code{results$desc} \tab \tab \tab \tab \tab a table \cr
  \code{results$fit} \tab \tab \tab \tab \tab a table \cr
  \code{results$est} \tab \tab \tab \tab \tab a table \cr
  \code{results$cp} \tab \tab \tab \tab \tab a table \cr
  \code{results$mem} \tab \tab \tab \tab \tab an output \cr
  \code{results$plot1} \tab \tab \tab \tab \tab an image \cr
  \code{results$plot} \tab \tab \tab \tab \tab an image \cr
}

Tables can be converted to data frames with \code{asDF} or \code{\link{as.data.frame}}. For example:

\code{results$mc$asDF}

\code{as.data.frame(results$mc)}
}
\description{
Latent Class Growth Modeling
}
