% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lltm.h.R
\name{lltm}
\alias{lltm}
\title{Linear Logisitic Test Model}
\usage{
lltm(
  data,
  vars,
  mat = "1,0,1,1,1,0,1,0,1,0",
  col = 2,
  items = TRUE,
  lr = FALSE,
  eta = FALSE,
  beta = TRUE,
  plot = TRUE,
  comp = TRUE
)
}
\arguments{
\item{data}{The data as a data frame.}

\item{vars}{.}

\item{mat}{.}

\item{col}{.}

\item{items}{.}

\item{lr}{.}

\item{eta}{.}

\item{beta}{.}

\item{plot}{.}

\item{comp}{.}
}
\value{
A results object containing:
\tabular{llllll}{
  \code{results$instructions} \tab \tab \tab \tab \tab a html \cr
  \code{results$text} \tab \tab \tab \tab \tab a preformatted \cr
  \code{results$items} \tab \tab \tab \tab \tab a table \cr
  \code{results$lr} \tab \tab \tab \tab \tab a table \cr
  \code{results$eta} \tab \tab \tab \tab \tab a table \cr
  \code{results$beta} \tab \tab \tab \tab \tab a table \cr
  \code{results$plot} \tab \tab \tab \tab \tab an image \cr
  \code{results$comp} \tab \tab \tab \tab \tab a table \cr
}

Tables can be converted to data frames with \code{asDF} or \code{\link{as.data.frame}}. For example:

\code{results$items$asDF}

\code{as.data.frame(results$items)}
}
\description{
Linear Logisitic Test Model
}
