% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lpa.h.R
\name{lpa}
\alias{lpa}
\title{Latent Profile Analysis}
\usage{
lpa(
  data,
  vars,
  nc = 2,
  variances = "equal",
  covariances = "zero",
  fit = TRUE,
  best = FALSE,
  est = FALSE,
  plot = FALSE,
  plot1 = FALSE,
  plot3 = FALSE,
  plot2 = FALSE,
  angle = 0,
  width = 500,
  height = 500,
  width1 = 500,
  height1 = 500,
  width2 = 500,
  height2 = 500,
  width3 = 500,
  height3 = 500
)
}
\arguments{
\item{data}{.}

\item{vars}{.}

\item{nc}{.}

\item{variances}{.}

\item{covariances}{.}

\item{fit}{.}

\item{best}{.}

\item{est}{.}

\item{plot}{.}

\item{plot1}{.}

\item{plot3}{.}

\item{plot2}{.}

\item{angle}{a number from 0 to 90 defining the angle of the x-axis labels,
where 0 degrees represents completely horizontal labels.}

\item{width}{.}

\item{height}{.}

\item{width1}{.}

\item{height1}{.}

\item{width2}{.}

\item{height2}{.}

\item{width3}{.}

\item{height3}{.}
}
\value{
A results object containing:
\tabular{llllll}{
  \code{results$instructions} \tab \tab \tab \tab \tab a html \cr
  \code{results$fit} \tab \tab \tab \tab \tab a table \cr
  \code{results$best} \tab \tab \tab \tab \tab a table \cr
  \code{results$est} \tab \tab \tab \tab \tab a table \cr
  \code{results$pc} \tab \tab \tab \tab \tab an output \cr
  \code{results$plot} \tab \tab \tab \tab \tab an image \cr
  \code{results$plot1} \tab \tab \tab \tab \tab an image \cr
  \code{results$plot3} \tab \tab \tab \tab \tab an image \cr
  \code{results$plot2} \tab \tab \tab \tab \tab an image \cr
}

Tables can be converted to data frames with \code{asDF} or \code{\link{as.data.frame}}. For example:

\code{results$fit$asDF}

\code{as.data.frame(results$fit)}
}
\description{
Latent Profile Analysis
}
