% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rasch.h.R
\name{rasch}
\alias{rasch}
\title{Rasch Model}
\usage{
rasch(
  data,
  vars,
  num = 1,
  step = 1,
  type = "RSM",
  aic = TRUE,
  bic = TRUE,
  caic = TRUE,
  imean = FALSE,
  imeasure = FALSE,
  ise = FALSE,
  infit = FALSE,
  outfit = FALSE,
  pbis = FALSE,
  wrightmap = FALSE,
  inplot = FALSE,
  outplot = FALSE,
  angle = 0,
  plot1 = FALSE
)
}
\arguments{
\item{data}{The data as a data frame.}

\item{vars}{.}

\item{num}{.}

\item{step}{.}

\item{type}{.}

\item{aic}{.}

\item{bic}{.}

\item{caic}{.}

\item{imean}{.}

\item{imeasure}{.}

\item{ise}{.}

\item{infit}{.}

\item{outfit}{.}

\item{pbis}{.}

\item{wrightmap}{.}

\item{inplot}{.}

\item{outplot}{.}

\item{angle}{a number from 0 to 45 defining the angle of the x-axis labels,
where 0 degrees represents completely horizontal labels.}

\item{plot1}{.}
}
\value{
A results object containing:
\tabular{llllll}{
  \code{results$instructions} \tab \tab \tab \tab \tab a html \cr
  \code{results$item$model} \tab \tab \tab \tab \tab a table \cr
  \code{results$item$items} \tab \tab \tab \tab \tab a table \cr
  \code{results$plot} \tab \tab \tab \tab \tab an image \cr
  \code{results$plot1} \tab \tab \tab \tab \tab an image \cr
  \code{results$inplot} \tab \tab \tab \tab \tab an image \cr
  \code{results$outplot} \tab \tab \tab \tab \tab an image \cr
  \code{results$total} \tab \tab \tab \tab \tab an output \cr
  \code{results$pmeasure} \tab \tab \tab \tab \tab an output \cr
  \code{results$pse} \tab \tab \tab \tab \tab an output \cr
  \code{results$pinfit} \tab \tab \tab \tab \tab an output \cr
  \code{results$poutfit} \tab \tab \tab \tab \tab an output \cr
}
}
\description{
Rasch Model
}
