% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rasch.h.R
\name{rasch}
\alias{rasch}
\title{Rasch Model}
\usage{
rasch(
  data,
  vars,
  num = 1,
  step = 1,
  type = "RSM",
  fit = TRUE,
  imean = TRUE,
  imeasure = TRUE,
  ise = FALSE,
  infit = FALSE,
  outfit = FALSE,
  pbis = FALSE,
  rsm = FALSE,
  pcm = FALSE,
  wrightmap = TRUE,
  lrsplit = "median",
  mlsplit = "median",
  waldsplit = "median",
  lr = FALSE,
  ml = FALSE,
  ml1 = FALSE,
  mlsplit1 = "median",
  wald = FALSE,
  inplot = FALSE,
  outplot = FALSE,
  angle = 0,
  plot1 = FALSE,
  plot2 = FALSE,
  plot3 = FALSE,
  piplot = FALSE
)
}
\arguments{
\item{data}{The data as a data frame.}

\item{vars}{.}

\item{num}{.}

\item{step}{.}

\item{type}{.}

\item{fit}{.}

\item{imean}{.}

\item{imeasure}{.}

\item{ise}{.}

\item{infit}{.}

\item{outfit}{.}

\item{pbis}{.}

\item{rsm}{.}

\item{pcm}{.}

\item{wrightmap}{.}

\item{lrsplit}{.}

\item{mlsplit}{.}

\item{waldsplit}{.}

\item{lr}{.}

\item{ml}{.}

\item{ml1}{.}

\item{mlsplit1}{.}

\item{wald}{.}

\item{inplot}{.}

\item{outplot}{.}

\item{angle}{a number from 0 to 45 defining the angle of the x-axis labels,
where 0 degrees represents completely horizontal labels.}

\item{plot1}{.}

\item{plot2}{.}

\item{plot3}{.}

\item{piplot}{.}
}
\value{
A results object containing:
\tabular{llllll}{
  \code{results$instructions} \tab \tab \tab \tab \tab a html \cr
  \code{results$item$model} \tab \tab \tab \tab \tab a table \cr
  \code{results$item$items} \tab \tab \tab \tab \tab a table \cr
  \code{results$rsm} \tab \tab \tab \tab \tab a table \cr
  \code{results$pcm} \tab \tab \tab \tab \tab a table \cr
  \code{results$plot} \tab \tab \tab \tab \tab an image \cr
  \code{results$piplot} \tab \tab \tab \tab \tab an image \cr
  \code{results$plot1} \tab \tab \tab \tab \tab an image \cr
  \code{results$plot2} \tab \tab \tab \tab \tab an image \cr
  \code{results$plot3} \tab \tab \tab \tab \tab an image \cr
  \code{results$inplot} \tab \tab \tab \tab \tab an image \cr
  \code{results$outplot} \tab \tab \tab \tab \tab an image \cr
  \code{results$total} \tab \tab \tab \tab \tab an output \cr
  \code{results$pmeasure} \tab \tab \tab \tab \tab an output \cr
  \code{results$pse} \tab \tab \tab \tab \tab an output \cr
  \code{results$pinfit} \tab \tab \tab \tab \tab an output \cr
  \code{results$poutfit} \tab \tab \tab \tab \tab an output \cr
  \code{results$lr} \tab \tab \tab \tab \tab a table \cr
  \code{results$ml} \tab \tab \tab \tab \tab a table \cr
  \code{results$ml1} \tab \tab \tab \tab \tab a table \cr
  \code{results$wald} \tab \tab \tab \tab \tab a table \cr
}

Tables can be converted to data frames with \code{asDF} or \code{\link{as.data.frame}}. For example:

\code{results$rsm$asDF}

\code{as.data.frame(results$rsm)}
}
\description{
Rasch Model
}
