% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rasch.h.R
\name{rasch}
\alias{rasch}
\title{Rasch Model}
\usage{
rasch(
  data,
  vars,
  num = 1,
  step = 1,
  type = "RSM",
  fit = TRUE,
  rel = TRUE,
  imean = TRUE,
  imeasure = TRUE,
  ise = TRUE,
  infit = FALSE,
  outfit = FALSE,
  pbis = FALSE,
  rsm = FALSE,
  pcm = FALSE,
  thr = FALSE,
  wrightmap = TRUE,
  lrsplit = "median",
  mlsplit = "median",
  waldsplit = "median",
  lr = FALSE,
  ml = FALSE,
  ml1 = FALSE,
  mlsplit1 = "median",
  wald = FALSE,
  gofplot = FALSE,
  tlab = "item",
  inplot = FALSE,
  outplot = FALSE,
  angle = 0,
  plot1 = FALSE,
  plot2 = FALSE,
  plot3 = FALSE,
  piplot = FALSE,
  plot4 = FALSE,
  plot5 = FALSE,
  res = "stdr",
  q3 = FALSE,
  res1 = "stdr",
  cormatrix = FALSE,
  width = 500,
  height = 500,
  width1 = 500,
  height1 = 500,
  width2 = 500,
  height2 = 500,
  width3 = 500,
  height3 = 500,
  width4 = 500,
  height4 = 500,
  width5 = 500,
  height5 = 500,
  width6 = 500,
  height6 = 500,
  width7 = 500,
  height7 = 500,
  plot8 = FALSE,
  num1 = 1,
  width8 = 500,
  height8 = 500,
  nptest = FALSE,
  matrix = 500,
  npmethod = "T11"
)
}
\arguments{
\item{data}{The data as a data frame.}

\item{vars}{.}

\item{num}{.}

\item{step}{.}

\item{type}{.}

\item{fit}{.}

\item{rel}{.}

\item{imean}{.}

\item{imeasure}{.}

\item{ise}{.}

\item{infit}{.}

\item{outfit}{.}

\item{pbis}{.}

\item{rsm}{.}

\item{pcm}{.}

\item{thr}{.}

\item{wrightmap}{.}

\item{lrsplit}{.}

\item{mlsplit}{.}

\item{waldsplit}{.}

\item{lr}{.}

\item{ml}{.}

\item{ml1}{.}

\item{mlsplit1}{.}

\item{wald}{.}

\item{gofplot}{.}

\item{tlab}{.}

\item{inplot}{.}

\item{outplot}{.}

\item{angle}{a number from 0 to 90 defining the angle of the x-axis labels,
where 0 degrees represents completely horizontal labels.}

\item{plot1}{.}

\item{plot2}{.}

\item{plot3}{.}

\item{piplot}{.}

\item{plot4}{.}

\item{plot5}{.}

\item{res}{.}

\item{q3}{.}

\item{res1}{.}

\item{cormatrix}{.}

\item{width}{.}

\item{height}{.}

\item{width1}{.}

\item{height1}{.}

\item{width2}{.}

\item{height2}{.}

\item{width3}{.}

\item{height3}{.}

\item{width4}{.}

\item{height4}{.}

\item{width5}{.}

\item{height5}{.}

\item{width6}{.}

\item{height6}{.}

\item{width7}{.}

\item{height7}{.}

\item{plot8}{.}

\item{num1}{.}

\item{width8}{.}

\item{height8}{.}

\item{nptest}{.}

\item{matrix}{.}

\item{npmethod}{.}
}
\value{
A results object containing:
\tabular{llllll}{
  \code{results$instructions} \tab \tab \tab \tab \tab a html \cr
  \code{results$items} \tab \tab \tab \tab \tab a table \cr
  \code{results$mf$model} \tab \tab \tab \tab \tab a table \cr
  \code{results$mf$rel} \tab \tab \tab \tab \tab a table \cr
  \code{results$thr} \tab \tab \tab \tab \tab a table \cr
  \code{results$rsm} \tab \tab \tab \tab \tab a table \cr
  \code{results$pcm} \tab \tab \tab \tab \tab a table \cr
  \code{results$plot} \tab \tab \tab \tab \tab an image \cr
  \code{results$piplot} \tab \tab \tab \tab \tab an image \cr
  \code{results$gofplot} \tab \tab \tab \tab \tab an image \cr
  \code{results$plot1} \tab \tab \tab \tab \tab an image \cr
  \code{results$plot2} \tab \tab \tab \tab \tab an image \cr
  \code{results$plot3} \tab \tab \tab \tab \tab an image \cr
  \code{results$inplot} \tab \tab \tab \tab \tab an image \cr
  \code{results$outplot} \tab \tab \tab \tab \tab an image \cr
  \code{results$ptotal} \tab \tab \tab \tab \tab an output \cr
  \code{results$pmeasure} \tab \tab \tab \tab \tab an output \cr
  \code{results$pse} \tab \tab \tab \tab \tab an output \cr
  \code{results$pinfit} \tab \tab \tab \tab \tab an output \cr
  \code{results$poutfit} \tab \tab \tab \tab \tab an output \cr
  \code{results$tm$lr} \tab \tab \tab \tab \tab a table \cr
  \code{results$tm$ml} \tab \tab \tab \tab \tab a table \cr
  \code{results$tm$ml1} \tab \tab \tab \tab \tab a table \cr
  \code{results$tm$wald} \tab \tab \tab \tab \tab a table \cr
  \code{results$plot4} \tab \tab \tab \tab \tab an image \cr
  \code{results$plot8} \tab \tab \tab \tab \tab an image \cr
  \code{results$plot5} \tab \tab \tab \tab \tab an image \cr
  \code{results$q3} \tab \tab \tab \tab \tab a table \cr
  \code{results$cormatrix} \tab \tab \tab \tab \tab a table \cr
  \code{results$text} \tab \tab \tab \tab \tab a preformatted \cr
  \code{results$text1} \tab \tab \tab \tab \tab a preformatted \cr
}

Tables can be converted to data frames with \code{asDF} or \code{\link{as.data.frame}}. For example:

\code{results$items$asDF}

\code{as.data.frame(results$items)}
}
\description{
Rasch Model
}
